restructured the ghis2s library to support various projects like HydroSFS. The application directory, s2s_app, is now removed from the ghis2s Python package which now sits disconnected from import ghis2s. S2Srun Class was moved to a new module main/experiment_setup
