+++ lis/utils/usaf/S2S/ghis2s/bcsd/metforce_temporal_disaggregation.py
@@ -62,7 +62,7 @@ def main(config_file, fcst_syr, fcst_eyr, month_abbr, month_num, job_name,
         config = yaml.safe_load(file)
 
     # Base forecast model
-    fcst_model = config['BCSD']['metforce_source']
+    fcst_model = config['BCSD']['source']['metforce']
 
     # get resolution
     lats, _ = utils.get_domain_info(config_file, coord=True)
@@ -73,7 +73,7 @@ def main(config_file, fcst_syr, fcst_eyr, month_abbr, month_num, job_name,
 
     lead_months = config['EXP']['lead_months']
     ens_num = config['BCSD']['nof_raw_ens']
-    model_name = config['BCSD']['metforce_source']
+    model_name = config['BCSD']['source']['metforce']
 
     # Path for where forecast files are located:
     forcedir = f"{projdir}/bcsd_fcst/{fcst_model}_{resol}"
