+++ lis/utils/usaf/S2S/ghis2s/bcsd/precip_temporal_disaggregation.py
@@ -64,7 +64,7 @@ def main(config_file, fcst_syr, fcst_eyr, month_abbr, month_num,
         config = yaml.safe_load(file)
 
     # Base forecast model
-    fcst_model = config['BCSD']['metforce_source']
+    fcst_model = config['BCSD']['source']['metforce']
 
     # get resolution
     lats, _ = utils.get_domain_info(config_file, coord=True)
