--- lis/utils/usaf/S2S/ghis2s/s2s_app/__init__.py
@@ -1,1 +1,0 @@
-from . import s2s_api
