+++ lis/utils/usaf/S2S/ghis2s/s2s_app/s2s_config_hydrosfs_fcast
@@ -5,12 +5,12 @@
 #######################################################################
 
 SETUP:
-  E2ESDIR: /discover/nobackup/projects/ghilis/S2S/GLOBAL/E2ES_557ww-7.5/
-  LISFDIR: /discover/nobackup/projects/ghilis/S2S/LISF/
+  E2ESDIR: /discover/nobackup/projects/ghilis/smahanam/SERVIR/SERVIR_5km/
+  LISFDIR: /discover/nobackup/projects/usaf_lis/smahanam/S2S/LISF-1/
   LISFMOD: lisf_7.5_intel_2023.2.1_s2s
   METFORC: /discover/nobackup/projects/usaf_lis/USAF_FORCING/
-  supplementarydir: /discover/nobackup/projects/usaf_lis/GHI_S2S/supplementary_files/
-  ldtinputfile: lis_input.s2s_global.noahmp401_hymap2.25km.nc
+  supplementarydir: /discover/nobackup/projects/ghilis/smahanam/SERVIR/supplementary_files/
+  ldtinputfile: lis_input.global_5km_hydroscs_mask_topo_merit1km_wclimppt.nc
   SPCODE:  s1189
   CONSTRAINT: mil
   DATATYPE: forecast
@@ -25,7 +25,7 @@ EXP:
   lsm:    noahmp401
   routing_name: hymap2
   pertmode: restart
-  lead_months: 9
+  lead_months: 4 
   NMME_models: [CanESM5, CESM1, CFSv2, GEOSv2, GFDL, GNEMO52]
 
 #######################################################################
@@ -33,14 +33,14 @@ EXP:
 #######################################################################
 
 BCSD:
-  fcst_download_dir: /discover/nobackup/projects/lis/MET_FORCING/CFSv2/
+  fcst_download_dir: /css/servir-s2s/HydroSFS/shared_data/GEOSv3/
   nmme_download_dir: /discover/nobackup/projects/usaf_lis/GHI_S2S/NMME/
-  clim_start_year: 1991
-  clim_end_year: 2020
-  nof_raw_ens: 12
-  source: {metforce: CFSv2, precip: nmme, obs: USAF-LIS7.3rc8}
-  force_corr: {lapsrate: false, aspect: false}
-
+  clim_start_year: 2002
+  clim_end_year: 2023
+  nof_raw_ens: 10
+  source: {metforce: GEOSv3, precip: null, obs: HydroSCS}
+  force_corr: {lapsrate: false, aspect: false}  
+  
 #######################################################################
 #                            LIS Forecast
 #######################################################################
@@ -48,8 +48,8 @@ BCSD:
 FCST:
   numprocx: 4
   numprocy: 72
-  JOB_SEGMENTS: {CanESM5: 3, CESM1: 3, GNEMO52: 3, GEOSv2: 2, CFSv2: 5,
-                 GFDL: 9}
+  JOB_SEGMENTS: {CanESM5: 2, CESM1: 2, GNEMO52: 2, GEOSv2: 2, CFSv2: 3,
+                 GFDL: 6}
   
 #######################################################################
 #                          Post-processor
@@ -65,4 +65,4 @@ POST:
   var_tair_min_list: [Tair_f_min]
   const_list: [lat, lon, ensemble, soil_layer, soil_layer_thickness, atime, Landmask_inst, LANDMASK, Landcover_inst, Soiltype_inst, Elevation_inst, Greenness_inst]
   metric_vars: [RZSM, TOP40SM, SFCSM, TWS, Precip, AirT, ET, Streamflow]
-  weekly_vars: [SWE, SnowDepth, SFCRELSM, SFCST, TOP40RELSM, TOP40ST, RZSM, Streamflow]
+  weekly_vars: [SWE, SnowDepth, RZSM, TOP40SM, TOP40ST]
